Refactor: Organize file structure
- I created a `documentation` directory to house project documentation.
- I moved `prd.md` and `project-plan.md` to the `documentation` directory.
- I deleted empty and unnecessary files: `Searching`, `Casks`, `Formulae`, and `src/index.css.backup`.
- I moved `src/lib/types.ts` to `src/types/` to centralize type definitions and updated all relevant imports.
